fix: move summary from da to abstract da
